+++ ai_bench/mlir/__init__.py
@@ -1,5 +1,5 @@
 from .compile import cpu_backend
-from .cpu_pipeline import cpu_pipeline
+from .pipeline import cpu_pipeline
 
 __all__ = [
     "cpu_backend",
